feat: Create separate module for exercises and solutions
