+++ app/src/main/kotlin/fr/o80/mostwanted/data/data.kt
@@ -11,17 +11,17 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import fr.o80.mostwanted.component.atom.toAnnotatedString
-import fr.o80.mostwanted.component.solution.S01HelloWorld
-import fr.o80.mostwanted.component.solution.S02Layout
-import fr.o80.mostwanted.component.solution.S03Screen
-import fr.o80.mostwanted.component.solution.S04Preview
-import fr.o80.mostwanted.component.solution.S05Image
-import fr.o80.mostwanted.component.solution.S06ClipCircle
-import fr.o80.mostwanted.component.solution.S07Interactive
-import fr.o80.mostwanted.component.solution.S08Animation
-import fr.o80.mostwanted.component.solution.S09Complex
-import fr.o80.mostwanted.component.solution.S09UsePreview
-import fr.o80.mostwanted.component.solution.S10TwoScreens
+import fr.o80.mostwanted.solution.S01HelloWorld
+import fr.o80.mostwanted.solution.S02Layout
+import fr.o80.mostwanted.solution.S03Screen
+import fr.o80.mostwanted.solution.S04Preview
+import fr.o80.mostwanted.solution.S05Image
+import fr.o80.mostwanted.solution.S06ClipCircle
+import fr.o80.mostwanted.solution.S07Interactive
+import fr.o80.mostwanted.solution.S08Animation
+import fr.o80.mostwanted.solution.S09Complex
+import fr.o80.mostwanted.solution.S09UsePreview
+import fr.o80.mostwanted.solution.S10TwoScreens
 import fr.o80.mostwanted.data.explanation.Explanation01
 import fr.o80.mostwanted.data.explanation.Explanation02
 import fr.o80.mostwanted.data.explanation.Explanation03
