+++ app/src/main/kotlin/fr/o80/mostwanted/exercises/ExerciseCard.kt
@@ -1,10 +1,10 @@
-package fr.o80.mostwanted.component
+package fr.o80.mostwanted.exercises
 
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Card
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -13,6 +13,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import fr.o80.mostwanted.ui.theme.WorkshopMostWantedTheme
 
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ExerciseCard(
     title: String,
@@ -20,7 +21,8 @@ fun ExerciseCard(
     modifier: Modifier = Modifier
 ) {
     Card(
-        modifier = modifier.clickable(onClick = onClick)
+        modifier = modifier,
+        onClick = onClick
     ) {
         Box(Modifier.padding(16.dp)) {
             Text(
