+++ build.gradle.kts
@@ -3,4 +3,5 @@ plugins {
     id("com.android.application") version "8.5.0" apply false
     id("org.jetbrains.kotlin.android") version "1.8.10" apply false
     id("com.google.dagger.hilt.android") version "2.47" apply false
+    alias(libs.plugins.android.library) apply false
 }
