+++ exercises/build.gradle.kts
@@ -1,22 +1,17 @@
 plugins {
-    id("com.android.application")
-    id("org.jetbrains.kotlin.android")
-    id("dagger.hilt.android.plugin")
-    kotlin("kapt")
+    alias(libs.plugins.android.library)
+    alias(libs.plugins.jetbrains.kotlin.android)
 }
 
 android {
-    namespace = "fr.o80.mostwanted"
+    namespace = "fr.o80.mostwanted.exercises"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "fr.o80.mostwanted"
         minSdk = 24
-        targetSdk = 34
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
         vectorDrawables {
             useSupportLibrary = true
         }
@@ -44,43 +39,21 @@ android {
     composeOptions {
         kotlinCompilerExtensionVersion = "1.4.3"
     }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
 }
-
-kapt {
-    correctErrorTypes = true
-}
-
 
 dependencies {
-    implementation(project(":exercises"))
-
-    implementation(libs.kotlinx.collections.immutable)
-
     implementation(libs.core.ktx)
     implementation(libs.lifecycle.runtime.ktx)
     implementation(libs.activity.compose)
     implementation(libs.androidx.compose.lifecycleruntime)
+    implementation(libs.androidx.compose.lifecycleruntime)
     implementation(platform("androidx.compose:compose-bom:2023.03.00"))
     implementation(libs.bundles.compose.ui)
-
-    implementation(libs.navigation)
-
-    kapt(libs.hilt.androidx.kapt)
-    kapt(libs.hilt.kapt)
-    implementation(libs.hilt.navigation)
-    implementation(libs.hilt.android)
 
     testImplementation(libs.junit)
 
     androidTestImplementation(libs.androidx.test)
     androidTestImplementation(libs.espresso.core)
-    androidTestImplementation(platform("androidx.compose:compose-bom:2023.03.00"))
-    androidTestImplementation(libs.androidx.compose.test)
 
     debugImplementation(libs.androidx.compose.ui.tooling)
     debugImplementation(libs.androidx.compose.ui.testmanifest)
