+++ exercises/src/main/kotlin/fr/o80/mostwanted/exercise/E01HelloWorld.kt
