+++ exercises/src/main/kotlin/fr/o80/mostwanted/exercise/E02Layout.kt
