+++ exercises/src/main/kotlin/fr/o80/mostwanted/exercise/E04Preview.kt
