+++ exercises/src/main/kotlin/fr/o80/mostwanted/exercise/E05to08Image.kt
