+++ exercises/src/main/kotlin/fr/o80/mostwanted/exercise/E10TwoScreens.kt
@@ -1,9 +1,8 @@
 package fr.o80.mostwanted.exercise
 
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 
 @Composable
-fun E10TwoScreens(modifier: Modifier = Modifier) {
+fun E10TwoScreens() {
 
 }
