+++ exercises/src/main/kotlin/fr/o80/mostwanted/internals/PreviewTheme.kt
@@ -1,0 +1,11 @@
+package fr.o80.mostwanted.internals
+
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+
+@Composable
+fun PreviewTheme(content: @Composable ()->Unit) {
+    MaterialTheme(
+        content = content
+    )
+}
