+++ exercises/src/main/kotlin/fr/o80/mostwanted/solution/S05Image.kt
@@ -1,4 +1,4 @@
-package fr.o80.mostwanted.component.solution
+package fr.o80.mostwanted.solution
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -11,8 +11,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import fr.o80.mostwanted.R
-import fr.o80.mostwanted.ui.theme.WorkshopMostWantedPreviewTheme
+import fr.o80.mostwanted.exercises.R
+import fr.o80.mostwanted.internals.PreviewTheme
 
 @Composable
 fun S05Image(
@@ -51,7 +51,7 @@ private fun PersonName(
 @Preview
 @Composable
 private fun S05ImagePreview() {
-    WorkshopMostWantedPreviewTheme {
+    PreviewTheme {
         S05Image()
     }
 }
