+++ exercises/src/main/kotlin/fr/o80/mostwanted/solution/S10TwoScreens.kt
@@ -1,4 +1,4 @@
-package fr.o80.mostwanted.component.solution
+package fr.o80.mostwanted.solution
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.core.tween
@@ -26,7 +26,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import fr.o80.mostwanted.ui.theme.WorkshopMostWantedPreviewTheme
+import fr.o80.mostwanted.internals.PreviewTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -97,7 +97,7 @@ fun Info(modifier: Modifier = Modifier) {
 @Preview
 @Composable
 private fun S10TwoScreensPreview() {
-    WorkshopMostWantedPreviewTheme {
+    PreviewTheme {
         S10TwoScreens()
     }
 }
