+++ exercises/src/main/res/drawable/avatar.xml
